requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
selenium>=4.0.0
webdriver-manager>=3.8.0
